Removed unnecessary code in `falcon_core`
